backend/pyproject.toml
[tool.poetry]
name = "valmed-backend"
version = "2.0.0"
description = "Professional medication logistics platform with Auth0 authentication and admin capabilities"
authors = ["ValMed Team"]
readme = "README.md"
packages = [{include = "backend", from = "."}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
pydantic-settings = "^2.1.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-json-logger = "^2.0.7"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
pydantic = {extras = ["email"], version = "^2.5.0"}
python-dotenv = "^1.0.0"
factory-boy = "^3.3.0"
faker = "^20.1.0"
requests = "^2.31.0"
python-keycloak = "^5.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
httpx = "^0.25.2"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
pytest-env = "^1.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn main:app --host 0.0.0.0 --port 8000"
dev = "uvicorn main:app --host 0.0.0.0 --port 8000 --reload"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto" 